pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "approvalml"
version = "0.1.0"
description = "YAML-based approval workflow parser and validator"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "hatchling",
]

[project.scripts]
approvalml = "cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
